implemented most of Dynamic and Fixed classes

+++ src/main/java/Entities/Bush.java
@@ -50,12 +50,12 @@ public class Bush extends Entity {
     }
 
     @Override
-    public int getHealth() {
+    public double getHealth() {
         throw new UnsupportedOperationException("Not supported yet.");
     }
 
     @Override
-    public int getEnergy() {
+    public double getEnergy() {
         throw new UnsupportedOperationException("Not supported yet.");
     }
