+++ src/main/java/Entities/Egg.java
@@ -22,12 +22,12 @@ public class Egg extends Entity {
      */
 
     @Override
-    public int getHealth() {
+    public double getHealth() {
         throw new UnsupportedOperationException("Not supported yet.");
     }
 
     @Override
-    public int getEnergy() {
+    public double getEnergy() {
         throw new UnsupportedOperationException("Not supported yet.");
     }
